Changes to handle interface role usage definition.
The previous implementation expected to mach each request to
in it's opposite interface role. The standard uses the roles
differently.

Received command messages should have the role of the receiver
and responses should have the role of the sender.

In the case of attribute responses no role validation is preformed,
as these commands can come from anywhere.