Fix issue with Registration messages.
If a registration message has the type of Protocol::Message::COMMAND_RESP_REQ,
then the Protocol::Filters::ResponseRequired filter will generate a
second message responding to the registration message even if the registration
response was already sent.

This is due to the fact that the source address for the incoming packet didn't
match the destination address for the response, given that this one already
had the newly allocated device address.