Refactor for consistency between Interfaces and Services.
This makes all base classes for the interfaces abstract and also makes it
so that the class that implements the 'uid' method both for services and
interfaces is called Base.

The helper template class for interfaces that implements the 'uid' method is
renamed to Interface to be consistent with implementing services using the
HF::Core::Service helper template.