[TESTS] FIX: Make tests pass after assert mock enabling.
Due enabling of the mock for the assert functionality a number of tests
need to be reviewed.

The tests that were checking for error conditions now hit the assert mock
and the expected asserts needed to be added.

For some tests in test_group_management.cpp, the assert mock enabling
reveled that the tests where actually not testing what was intended.