FIX: Interface attribute wrapper bad size calculation
When creating a attribute API wrapper for an Interface object using
getter/setter functions the size function was not updating the value
used by the helper object to calculate the size thus providing invalid
sizes for pack/unpack of the attribute.

In order to maintain the correct synchronization of the attribute with
the underlying value, recreate the helper object every time the
serialization/compare API is used.